Convert hipchat client in a puller
Will create an struct for each token and channel to listen (using
pull method). If two or more groups share the same hipchat channel,
the bot will pull this channel just one time.

On commmand /ping or /catch action the confirmation is sent by
each group, so we will se the same message more than one time in
the hipchat or telegram room. Implemented a simple cache to don't
send the same message to the same channel in short window time.

Code organization